pyproject.toml
[project]
name = "docscanai"
version = "0.1.0"
description = "Local-first document scanning, OCR, and structured extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pymupdf>=1.25",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.7",
    "opencv-python-headless>=4.13.0.92",
    "pytesseract>=0.3.13",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"backend/config.py" = ["N802"]
